Try and fix typings
